requirements.txt
pandas>=2.1.3
scikit-learn>=1.3.0
pyyaml>=6.0.0
python-dateutil>=2.8.2
numpy>=1.26.0
joblib~=1.3.2
xgboost~=2.0.2
dvc>=3.30.3
pydantic>=2.5.0